#00 [Chore] 기타정보 horizontal recyclerview로 변경 및 패키지로 분리
