+++ app/src/main/java/com/thequietz/travelog/guide/model/RecommendResponse.kt
