+++ app/src/main/java/com/thequietz/travelog/guide/viewmodel/OtherViewModel.kt
@@ -1,10 +1,12 @@
-package com.thequietz.travelog.guide
+package com.thequietz.travelog.guide.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.thequietz.travelog.data.RepositoryImpl
+import com.thequietz.travelog.guide.Place
+import com.thequietz.travelog.guide.RecommendPlace
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
