+++ app/src/main/res/layout/fragment_other_info.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.thequietz.travelog.guide.OtherViewModel" />
+            type="com.thequietz.travelog.guide.viewmodel.OtherViewModel" />
     </data>
     <ScrollView
         android:layout_width="match_parent"
@@ -16,7 +16,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".guide.OtherInfoFragment">
+        tools:context=".guide.view.OtherInfoFragment">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/tb_other_info"
@@ -94,9 +94,8 @@
             android:layout_marginEnd="20dp"
             app:layout_constraintTop_toBottomOf="@id/tv_vacation_spot"
             android:layout_marginTop="10dp"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-            app:spanCount="2"
-            android:orientation="vertical"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            android:orientation="horizontal"
             android:id="@+id/rv_vacation_spot"
             />
         <TextView
@@ -133,9 +132,8 @@
             android:layout_marginTop="10dp"
             app:layout_constraintTop_toBottomOf="@id/tv_food"
             android:id="@+id/rv_food"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-            app:spanCount="2"
-            android:orientation="vertical"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            android:orientation="horizontal"
             />
         <TextView
             android:layout_width="wrap_content"
@@ -170,9 +168,8 @@
             android:layout_marginEnd="20dp"
             android:layout_marginTop="10dp"
             app:layout_constraintTop_toBottomOf="@id/tv_fesetival"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
-            app:spanCount="2"
-            android:orientation="vertical"
+            app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+            android:orientation="horizontal"
             android:id="@+id/rv_festival"
             />
