+++ app/src/main/res/layout/fragment_specific_guide.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.thequietz.travelog.guide.SpecificGuideViewModel" />
+            type="com.thequietz.travelog.guide.viewmodel.SpecificGuideViewModel" />
         <import type="android.view.View"/>
     </data>
 
@@ -18,7 +18,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="@color/white"
-            tools:context=".guide.SpecificGuideFragment">
+            tools:context=".guide.view.SpecificGuideFragment">
 
             <TextView
                 android:layout_width="wrap_content"
