+++ app/src/main/res/layout/item_recycler_other_info.xml
@@ -12,7 +12,8 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:paddingBottom="20dp">
+        android:paddingBottom="20dp"
+        android:paddingEnd="20dp">
         <ImageView
             android:layout_width="170dp"
             android:layout_height="150dp"
