+++ app/src/main/res/navigation/record.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/record"
-    app:startDestination="@id/recordFragment">
+    app:startDestination="@id/recordViewOneFragment">
 
     <fragment
         android:id="@+id/recordFragment"
